Create ra module
